Fix lifetime encoding

+++ NET/LibUA/Types.cs
@@ -6887,7 +6887,7 @@ namespace LibUA
 
             public uint ChannelID { get; set; }
             public uint TokenID { get; set; }
-            public UInt32 TokenLifetime { get; set; }
+            public double TokenLifetime { get; set; }
             public DateTimeOffset TokenCreatedAt { get; set; }
 
             public uint? PrevChannelID { get; set; }
